requirements.txt
Flask
pillow
tensorflow